// Final_FrontEnd_Angular/src/app/body/api/exam/Biometry.ts
export interface Biometry {
    id: number;
    createdBy: string;
    createdAt: string;
    modifiedBy: string;
    modifiedAt: string;
    deleted: boolean;
    version: number;

    eyeSide: 'OU' | 'OD' | 'OS' | null;
    examType: string | null;
    examDate: string | null;
    examComment: string | null;
    examQuality: string | null;
    calculDate: string | null;
    eyeStatus: string | null;
    selected: boolean;
    al: number | null;
    acd: number | null;
    internalAcd: number | null;
    pupilDia: number | null;
    pupilMin: number | null;
    pupilMax: number | null;
    wtw: number | null;
    cord: number | null;
    z40: number | null;
    hoa: number | null;
    kappaAngle: number | null;
    n: number | null;
    k1: number | null;
    k1Mm: number | null;
    k1Axis: number | null;
    k2: number | null;
    k2Mm: number | null;
    k2Axis: number | null;
    kAstig: number | null;
    kAstigAxis: number | null;
    kAvg: number | null;
    rAvg: number | null;
    k1CorneaBack: number | null;
    k2CorneaBack: number | null;
    k1AxisCorneaBack: number | null;
    k2AxisCorneaBack: number | null;
    siaCyl: number | null;
    siaAxis: number | null;
    snr: number | null;
    lensThickness: number | null;
    cct: number | null;
    cctMin: number | null;
    asphQf: number | null;
    asphQb: number | null;
    alStatus: number | null;
    alError: number | null;
    kPreRefrAvg: number | null;
    rPreRefrAvg: number | null;
    manifestRefrDate: string | null;
    manifestRefrSph: number | null;
    manifestRefrCyl: number | null;
    manifestRefrAxis: number | null;
    manifestRefrVd: number | null;
    targetRefrSph: number | null;
    targetRefrCyl: number | null;
    machine: string | null;
    importType: string | null;
    patientId: number | null;
}